Variable names are standardized for easier maintenance in the future.
The current code uses "$data" and "default_sec" for both "rc" and "ini" files, sometimes redeclaring in the same scope.

Two types of prefixes are used to contextually identify the variables: "rc" and "ini_".

The following changes are made:

* `$rcfile` -> `$rc_file` (same style as `$ini_file`)
* `$data` -> `$rc_data`, `$ini_data`
* `$default_sec` -> `$rc_default_sec`, `$ini_default_sec`

This should make it much easier for contributors to review code.